pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-food-delivery-env-v2"
version = "0.1.0"
description = "Food delivery dispatch environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.7.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
train = [
    "torch>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["food_delivery_env_v2", "food_delivery_env_v2.server"]
package-dir = { "food_delivery_env_v2" = ".", "food_delivery_env_v2.server" = "server" }
